+++ ts/packages/providers/vercel/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @composio/vercel
 
+## 0.1.31
+
+### Patch Changes
+
+- e660385: Fix file upload / download
+- Updated dependencies [e660385]
+  - @composio/core@0.1.31
+
 ## 0.1.30
 
 ### Patch Changes